requirements.txt
# --- Core Data Processing ---
numpy==2.0.1
pandas==2.3.2
pytz==2025.2
tqdm==4.67.1

# --- Vision & Image Processing ---
opencv-python-headless==4.10.0.84
scikit-image>=0.24.0
pillow>=10.4.0

# --- AI Framework (PyTorch Ecosystem) ---
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
jinja2==3.1.4

# --- Object Detection (YOLO) ---
ultralytics==8.3.40
thop==0.1.1.post2209072238

# --- Machine Learning Tools ---
scikit-learn==1.5.2
joblib==1.5.2
ttach==0.0.3

# --- Visualization & Explainability ---
matplotlib==3.10.6
seaborn==0.13.2
grad-cam==1.5.5
pytorch-gradcam==0.2.1

# --- API & Cloud Integration ---
requests>=2.31.0
PyYAML>=6.0.0
google-genai==1.34.0
google-auth==2.40.3

# --- Experiment Tracking ---
wandb==0.22.2
tensorboard==2.18.0
protobuf==4.25.8

# --- Testing ---
pytest==8.4.2